autodoc_1.0/requirements.txt
openai
astunparse
GitPython